DPAA/NetCommSw: make irq flags variables unsigned long
local_irq_save()/restore() really should use unsigned long
and not just uint32_t. Fix that.

+++ drivers/net/dpa/NetCommSw/Peripherals/FM/HC/hc.c
@@ -86,7 +86,7 @@ do {                                                                            
 
 #define TRY_LOCK                                                                \
 do {                                                                            \
-    uint32_t intFlags;                                                          \
+    unsigned long intFlags;                                                          \
     intFlags = XX_DisableAllIntr();                                             \
     if (p_FmHc->lock)                                                           \
     {                                                                           \
@@ -99,7 +99,7 @@ do {                                                                            
 
 #define TRY_LOCK_RETURN_NULL                                                    \
 do {                                                                            \
-    uint32_t intFlags;                                                          \
+    unsigned long intFlags;                                                          \
     intFlags = XX_DisableAllIntr();                                             \
     if (p_FmHc->lock)                                                           \
     {                                                                           \
